tests/wpt/tests/html/obsolete/requirements-for-implementations/the-marquee-element-0/marquee-scrollamount.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: marquee-scrollamount</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="https://html.spec.whatwg.org/multipage/multipage/obsolete.html#the-marquee-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<marquee id="test1" scrollamount="aa">Test1</marquee>
<marquee id="test2" scrollamount="-1">Test2</marquee>
<marquee id="test3" scrollamount="10">Test3</marquee>
<script>
  test(function() {
    var mq = document.getElementById("test1");
    assert_equals(mq.scrollAmount, 6, "The value of scrollamount should be 6.");
  }, "The scrollamount is a string");

  test(function() {
    var mq = document.getElementById("test2");
    assert_equals(mq.scrollAmount, 6, "The value of scrollamount should be 6.");
  }, "The scrollamount is a negative");

  test(function() {
    var mq = document.getElementById("test3");
    assert_equals(mq.scrollAmount, 10, "The value of scrollamount should be 10.");
  }, "The scrollamount is a normal value");
</script>
